chore(ui-components): add generated files for Notebookitem, PlanCard, ReacticonsbiBiNote, and Sidebar
feat(ui-components): add Sidebar component

This commit adds a new file, Sidebar.jsx, which contains the implementation of the Sidebar component. The contents of this file were generated with Amplify Studio and should not be modified. Any changes made to this file will be overwritten when running amplify pull. The Sidebar component is a React component that renders a sidebar with navigation items and buttons. It includes icons from the react-icons library and uses the @aws-amplify/ui-react package for UI components. The Sidebar component also includes a footer section with support links and a logo.